version 4.1
February, 3th 2008

    When several artifacts have the same name and you edit a class, the shown artifact may not the right one, and if you change it the class may not be associated to the desired one, fixed.

    New drawing setting named show members multiplicity allowing to write attribute multiplicity in a class picture in a class diagram. This setting have sense only when you also ask for to show the full members definition and when the drawing language is UML. By default this setting is set to false.

    New drawing setting named max members width allowing to limit attribute and operation width to a given number of characters. When the number of characters needed by an attribute or an operation is greater than max members width, the string is cut to max members width characters and "..." is added. Note : the optionally written visibility and stereotype are not taken into in the member width.

    XMI2 Import 1.2.5
    Spaces are replaced by '_' in the name of an imported class, attribute or relation.
    Fixe bug on operation parameter multiplicity reading.

    Because the format of the BOUML files is changed, the previous releases of BOUML cannot read a project saved with this version. Obviously this release is able to read the projects made by previous releases of BOUML.

+++ src/RoundtripBody/UmlOperation.cpp
@@ -137,7 +137,7 @@ void UmlOperation::roundtrip(const char * path, aLanguage who)
       if (op == 0) {
 	QCString n;
 	
-	n.sprintf("%x", id);
+	n.sprintf("%x", (unsigned) id);
 	UmlCom::trace(QCString("<font  color =\"red\"> Error in ") + path + 
 		      " : unvalid operation id " + n + "</font><br>");
 	UmlCom::bye();
